:enhance receive kernel version as param

+++ build_kernel_from_src.sh
@@ -4,22 +4,30 @@ TEMP_DIR=/home/pi/temp
 mkdir $TEMP_DIR
 cd $TEMP_DIR
 
-SOURCE_URL=https://ghproxy.com/https://github.com/raspberrypi/linux/archive/refs/heads/rpi-5.15.y.zip
+if [ -z "$1" ]
+then
+    version=5.5 #default
+else
+    version=$1
+fi
 
-wget $SOURCE_URL
-
+SOURCE_URL=https://ghproxy.com/https://github.com/raspberrypi/linux/archive/refs/heads/rpi-$version.y.zip
 zip_name=$(basename $SOURCE_URL)
+unziped_folder_name=linux-$(basename $zip_name .zip)
 
-unzip $zip_name
-rm $zip_name
+if test -e $unziped_folder_name; then
+    cd $unziped_folder_name
+else
+    wget $SOURCE_URL
+    unzip $zip_name
+    rm $zip_name
+    cd $unziped_folder_name
+fi
 
-unziped_folder_name=$(basename $zip_name .zip)
-cd $unziped_folder_name
-
+echo "WORK DIR: $(pwd)"
 KERNEL=kernel7l
-make bcm2711_defconfig
+sudo make bcm2711_defconfig
 
-CONFIG_LOCALVERSION="-v7l-MY_CUSTOM_KERNEL"
 
 make -j4 zImage modules dtbs
 sudo make modules_install
@@ -28,4 +36,10 @@ sudo cp arch/arm/boot/dts/overlays/*.dtb* /boot/overlays/
 sudo cp arch/arm/boot/dts/overlays/README /boot/overlays/
 sudo cp arch/arm/boot/zImage /boot/$KERNEL.img
 
-sudo apt install raspberrypi-kernel-headers
+
+<<COMMENT
+headers_dir=/usr/src/linux-headers-$(uname -r)
+sudo mkdir $headers_dir
+sudo cp * $headers_dir -r
+sudo ln -s $headers_dir /lib/modules/$(uname -r)/build
+COMMENT
